Fix self-hosted build error on macOS with Qt 5.7
This was caused by a change in Qt 5.7 which #includes the Availability.h
header which is not in moc's default search path. Qbs 1.6 contains a
patch to supply moc with the correct header search paths, therefore this
workaround is required until Qbs requires version 1.6 of itself for self
hosted builds.

Change-Id: I38db36235ac34a06a1a8e33fb474bb72b2928184
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>