Introduce the --wait-lock command line option
This allows multiple qbs processes competing for the same configuration
instance to be launched simultaneously, and the inactive instances will
wait indefinitely for the lock file to be released.

This is primarily intended for use by the IDE generator plugins like
Visual Studio where the "Build Solution" command builds all projects in
the solution in parallel (and this cannot be disabled).

Change-Id: I13c37a21dd8386e3dfdab1f769b8242bfcfc2e45
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>