Introduce cpp.linkerMode
This property allows qbs to automatically select an appropriate linker
based on the objects being linked. The new behavior is to automatically
use the C++ compiler driver for linking if any of the objects being
linked were compiled from C++ or Objective-C++ sources, otherwise the C
compiler driver if any of the objects being linked were compiled from C
or Objective-C sources, otherwise the system linker. To retain the old
behavior, cpp.linkerName can be set to cpp.cxxCompilerName, and
cpp.linkerMode to "manual". Qbs will also avoid passing the -stdlib
option or linking c++abi on Linux unless any of the objects being linked
were compiled from C++ or Objective-C++ sources.

As a result the default behavior for building a pure C application or
pure ASM application with Qbs is to NOT link to the C++ standard library
or any other supporting C++ libraries.

[ChangeLog] Introduced cpp.linkerMode property to allow selection of the
correct linker (C driver, C++ driver, or system linker) based on the
objects being linked

Change-Id: Ic16e31a3072c9e296c348783a21717902c15bff6
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>