Separate file scope and imports scope
When evaluating, imported names need to have a higher precedence than
names in the item scope, as otherwise they can clash with dependencies
of the product (see bug report and autotest). We cannot simply move the
file scope's precedence up, because in addition to imports and the
"file" and "filePath" properties, it also potentially contains ids from
e.g. a module prototype, which would then override the respective entry
in the module scope.

Task-number: QBS-930
Change-Id: Ifb8b7c933225f872ccc3e878b2bf01b7b0ac0f99
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>