Fix change tracking for precompiled headers
We used "explicitlyDependsOn" in the respective rules to make the pch
sources depend on all header files. This was much too coarse, as it
meant that the precompiled header was rebuilt whenever any header file
in the product changed. Instead, we need to scan the pch sources just
like all the other C and C++ source files to find out which header files
they actually depend on.

Change-Id: Iaec4ed85075b387d9f4b0fcd06b1e481851c6b10
Reviewed-by: Jake Petroules <jake.petroules@qt.io>