ModUtils: Filter out empty strings from string lists
It can easily happen due to subtle user errors that an empty string ends
up in an array. The resulting error message is often rather baffling to
users. Amend the convenience functions in ModUtils so that empty strings
are removed from the values of list properties.
Note that we do not do this in qbscore because it is not technically
wrong to have an empty string in an array. Module authors who want to
support such properties can opt out of the filtering by using
{product,artifact}.moduleProperty() directly, rather than the
convenience functions from ModUtils.

Task-number: QBS-873
Change-Id: Ie752f14b3625287f846d945e548920181186f22f
Reviewed-by: Jake Petroules <jake.petroules@qt.io>