More sensible fallback values for "base" & friends
When setting up the values of special properties like "base", "outer"
and "original", we replaced invalid and undefined values with an empty
array. This is all nice and well for array properties (so that people
can write "base.concat()" without checking base first), but quite absurd
for scalar properties. So for the latter, the only thing we do now is
replace invalid values with undefined ones. The behavior for arrays
stays the same.

Change-Id: Icb3c74927c84025a07e3ef6c9c67b89d95320632
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>