Do not needlessly re-evaluate module properties in Groups
For a Group that contains at least one module property assignment, we
used to re-evaluate all module properties. This was of course immensely
wasteful, as there are normally only a handful of properties in need of
re-evaluation.
We fix the problem in the following way:
    - When evaluating a product's module properties, keep track of
      the dependencies between them.
    - For all Groups of that product, re-evaluate only the properties
      which are either explicitly set in that group or have a dependency
      on such a property.

Here's the output of qbs_benchmarker, using qbs itself as the test
project:
========== Performance data for Resolving ==========
    Old instruction count: 2083482809
    New instruction count: 1696643887
    Relative change: -19 %
    Old peak memory usage: 2997349 Bytes
    New peak memory usage: 2704033 Bytes
    Relative change: -10 %

I've also profiled with the Qt Creator super project (not using
valgrind, as that would take ages). For that project, this change halves
the time spent in the ProjectResolver, while the overall resolving time
goes down by about a third.

Note that we could potentially use the gathered dependency data in
change tracking as well, provided we manage to find out what exactly
changed in the project files. That should result in an improved
experience for IDE users.

Change-Id: I6ceffe7a6f0421617f53a4dd2eade51993c2785f
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>