Change the default base directory for entries of the "files" list
It used to be the product source directory, but that is unintuitive and
usually not what the user wants when groups are pulled in from other
files such as modules.
Also add documentation for this, which was missing altogether.

[ChangeLog] The base directory for source files changed from the product
source directory to the parent directory of the file where the files are
listed.

Task-number: QBS-849
Change-Id: Ia424eab52adba8b3aed1e482a0e081fa35fb6429
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>