GCC module: Do not add assume libc++abi with libc++ on Linux
We hardcoded the respective flag because it was observed that linking
failed without it. However, this requirement seems to be a
(questionable) peculiarity of one specific Linux distribution (namely
ArchLinux). In general, one should not need to add the low-level ABI
library. Where it is required, the user can add it to the list of
libraries (possibly via a profile).

Task-number: QBS-878
Change-Id: Iece8e1af566b0696419b84c82d076841e1cb9c66
Reviewed-by: Jake Petroules <jake.petroules@qt.io>