Change the default value and behavior of Qt.core.resourceSourceBase
The default value "undefined" tells the qrc file generator to use
only the input file name as a resource alias.

The benefits of changing the default value are as follows:
1) It's a bit weird to reset a normal value to undefined in qbs code.
2) This simplifies feature usage for output Artifacts and single files.
3) If one wants to put a hierarchical structure into a qrc file
there are a lot of cases in which they have to specify
resourceSourceBase manually.
4) When a Group with a wildcard and a prefix is used this makes user's
intention a bit clearer. This strongly separates "I want this set of
files" from "I want this directory structure".

Task-number: QBS-1053
Change-Id: Ib6413643398dca475b4b71ad884202c9e7fb823a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>