cpp module: Consider only absolute file paths as paths to external libs
In properties like cpp.dynamicLibraries, users can give names ("m") or
actual file paths ("/opt/lib/mylib.so"). In the first case, we prepend
the -l option, whereas in the second case the name is added to the
command line as-is. In the past, we have used different heuristics to
determine whether the name is a path:
    - In the beginning, it was dark and cold, and I don't remember what
      we did.
    - Later, a heuristic was introduced that tried to determine from the
      string whether it was a file name. This resulted in false
      negatives -- we encountered libraries that had unexpected
      names.
    - To fix the problem of the approach above, we then checked whether
      the given string existed as a file. This resulted in false
      positives -- once in a while, a string denoting a library name
      matched a file name in the user's working directory. The shorter
      the library name, the higher the likelihood of that happening.
So let's just check whether the string is an absolute path. That
unambiguously tells us that we are dealing with an actual file. I don't
see a good reason for relative file paths to occur in these properties.

Task-number: QBS-1047
Change-Id: I7d0109523af1fca4bc2f8bee5e34c9ecc199d35a
Reviewed-by: Jake Petroules <jake.petroules@qt.io>