Change style of #includes of Qt headers
By using q<lowercase>.h headers, we become more bootstrap-friendly, as no
full syncqt is required any more.

In the same go, prefix all includes with the module name. This helps
grep-based analysis of which parts of Qt are used and where. It's also
consistent with Qt's public headers (where syncqt enforces the style).

Testdata and examples are excluded from the change, as they are not
relevant for bootstrapping.

Change-Id: I9539b1d795e87fca6e5fc6c91acbb775b79208d9
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>