Make Win32 API usage long-path aware
Most Win32 API functions are limited to working with paths which are
up to MAX_PATH (260) characters long by default. Windows 10 version 1607
allows this limit to be exceeded by setting a LongPathsEnabled registry
key and manifesting the application with the longPathAware=true setting.
We cannot rely on this being the case since we support older Windows
versions and the registry key is set to off by default.

Fortunately, most Win32 API functions allow prepending the \\?\ prefix
which raises the path length limit to UNICODE_STRING_MAX_CHARS (32767)
characters. This works on all Windows versions.

For the FileInfo class in particular, this resolves potential false
negatives for file existence, inability to retrieve modification times,
and misidentification of directories as files.

Task-number: QBS-1068
Change-Id: Ia4a37fa75ede803e3e2ecb63ae1ab12d2dae33bd
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>