Downgrade error message from JavaCompilerScanner to a warning
This is not actually fatal, as Qbs collects dependency information both
from the TreePathScanner API *and* the output files the compilation
process writes to. Instead, downgrade it to a warning and also point
users to the bug report. This fixes a regression from 1.5.x.

Adds an XFAIL'ing autotest.

Task-number: QBS-1069
Change-Id: I30e5f5d0d1554bf5078509768fdad12f20922319
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>