Fix some issues in the WiX module causing the autotest to fail
This allows qbs.architecture to be undefined for WiX, in which case the
WiX default architecture of x86 will be used. The autotest would fail if
there was no architecture set in the profile (which is allowed).

This also corrects the registry key name for finding WiX, leading to
incorrect autotest failures if WiX tools were in not in the PATH or
WiX properties were not set in a profile.

Lastly, this fixes a wrong concatenation of product.buildDirectory and
improves the output and error reporting for the autotest.

Change-Id: I227924b681131c13fcb09d2d5010b372ff44b241
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>