Merge the C/C++ include scanners
We had one scanner for every type of file in the C family, all of which
collected the same set of includes, but under a different entry in the
scanner cache. Thus, lots of header files were unnecessarily re-scanned.
We fix this by making the scanner plugins declare a *list* of tags they
can handle and passing the currently active tags in the open() function.

========== Performance data for Rule Execution ==========
    Old instruction count: 3775190973
    New instruction count: 3441085735
    Relative change: -9 %
    Old peak memory usage: 18174216 Bytes
    New peak memory usage: 17307600 Bytes
    Relative change: -5 %

Change-Id: I222d1ec4bbfbc06ecd8c81faa55a500bc0da1ee6
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>