Change tracking: Do not re-expand wildcards if we know we don't have to
We now remember the time stamps of all directories in which we looked
for wildcards. If none of these time stamps have changed, then there can
be no new wildcard matches and we don't have to do the re-expansion.
The output of our benchmarking tool is below. The real-world savings
appear to be higher, though.

========== Performance data for Null Build ==========
    Old instruction count: 461543852
    New instruction count: 440917216
    Relative change: -5 %
    Old peak memory usage: 13134440 Bytes
    New peak memory usage: 13108120 Bytes
    Relative change: -1 %

Change-Id: I8b126786f8b5ccffa70a6a5e64c27aae8d23478b
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>