Add support for the Qt Quick compiler
Detect the availability of the Qt Quick compiler in qbs-setup-qt.
Add a property bool Qt.quick.compilerAvailable that is true if the Qt
Quick compiler is available in this Qt build.

If the compiler is available, create an extended Qt.quick module with
Qt Quick compiler rules. Create a standard module otherwise. The rules
pick up Qt resource files tagged with Qt.quick.qrc and compile the
referenced QML files. The remaining files are put into a new QRC file
that is then passed to rcc.

A prioritized file tagger is provided that automatically tags *.qrc
files as "qt.quick.qrc" if the Qt quick compiler is available.

[ChangeLog] Added support for the Qt Quick compiler.

Task-number: QBS-749
Change-Id: I44e490e09796afeb5455ec50abc1e798f7eaf01a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>