Do not store undefined path values as null strings in property map
The path property conversion in the ProjectResolver turned every
undefined path property into a null string. That was the reason for the
isNull() check in the PropertyFinder. That check can now be removed.

Note that we still run the conversion code for string lists and path
lists, because currently the expected outcome is an empty array in both
cases.

Change-Id: I2695b86786dc42319afbf7a896dccc02979e094a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>