Fix restoration of symbolic commits
qbs-benchmarker tries to restore the original commit of the qbs
repository. To determine that commit the command
    git describe HEAD
is used. For the 1.7 branch this produces something like
    v1.7.0-26-g2bc9373
which then is passed to 'git checkout'.
Afterwards the repo is in detached HEAD state.

Use 'git symbolic-ref' to determine the correct original branch name.
Fall back to the old method if that doesn't work.

Change-Id: I21526db1efceebcf5102eed9da43478871880ab1
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>