Make loaded extensions available in JavaScript commands
When JS commands are created in a function that's defined in a separate
JS file then loaded extensions of this file were not visible within the
JS command's source code.

Determine the current import scope name at the location of the
JavaScriptCommand's constructor and pass it to the JsCommandExecutor
where it is pushed to the current context.

Task-number: QBS-1093
Change-Id: I471e277c6821f38af6f2a06fc34005cf3c296f6c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>