Reduce overcautious rule connections
Before, we connected all leaf rules of a product with all root rules of
its product dependencies. This is not necessary and penalizes parallel
builds of dependent products.

Now, we connect all rules of a product with all compatible rules of
product dependencies. Rules that take "installable" as input still use
the old behavior, because "installable" is a magic tag with no
corresponding rule.

Change-Id: Ib1f8b8854bf9d265c155f0d34d321c0e10607c1d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>