RulesApplicator: fix bug with setting output artifact module properties
This patch fixes module property extraction from the output artifacts of
a Rule. Previously, a variable was inappropriately modified in a loop
which caused an output artifact which applied module properties to two
modules, for example, 'cpp' and 'bundle', to attempt to apply properties
to an erroneous module named 'cpp.bundle'.

Change-Id: If20851f3d6d276aff15f11446e12b52c2a0410db
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>