Match a rule's input tags against root artifacts of dependencies
For instance, if you make an object file the target artifact of your
product, you want the depending product's linker rule to pick it up,
without the need to explicitly add the tag to "inputsFromDependencies".
Otherwise, you'd basically have to duplicate the contents of "inputs" in
all rules then, which is silly.

[ChangeLog] A rule's "inputs" property is now matched against the file
tags of target artifacts of products that the current product depends
on, in addition to artifacts in the current product.

Change-Id: I814cb49f5d8135b7482777dce73f1fa56a44724b
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>