Handle single-architecture Android builds more gracefully
We only set a default list of architectures to compile for, if the
current profile does not have a qbs.architecture property set (as should
be the case for a top-level Android profile). Otherwise, the
architectures list is set to undefined, which allows the Product to use
the default architecture set in the profile.

This allows products which do not touch the architectures property to
work in Qt Creator out of the box by default, until such time as
Qt Creator adds support for multi-arch Android building.

Change-Id: Iaaab5e10b025aa1504e9ce9dbe28d41f7148bfb0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>