Do not pass --undefined-symbols=ignore-in-shared-libs
Now that we correctly add -rpath-link to the linker command line we want
errors regarding missing symbols to be reported. This patch removes the
--undefined-symbols=ignore-in-shared-libs argument that was added in
commit d9127c78.

Since ld's man page claims that report-error is the default, do not pass
an --undefined-symbols=XXX argument. On Darwin the behavior is not
changed.

Interestingly, passing --undefined-symbols=report-all on FreeBSD will
result in the weird linker errors mentioned in commit d9127c78.

The libc on FreeBSD refers to symbols that only occur in applications,
such as "__progname" and "environ". Thus, no dynamic library can be
linked with "report-all". The ld documentation is not correct in claiming
that this option is the default; the linker clearly does some additional
magic, like blacklisting some symbols or whatever.

Change-Id: I330b8ab7dd9fbf777946992914239cce5a43fd8d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>