GCC: Do not mess with the list of external libraries
Consider the following:
There are external static libraries libA and libB, with libB having a
dependency on libA. Some module pulls in libA, another one pulls in libB
and libA.
The raw contents of cpp.staticLibraries for the product might
end up as ["libA","libB","libA"]. Before this patch, our library
collection code "simplified" this list to ["libA","libB"], which leads
to a linker error. Since we generally do not know about the dependencies
of external libraries, we should not touch these lists.

Change-Id: I4a94dfc8505ee9ece2562dc3aa3d66244ec7fdac
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>