Improve dependency tracking between module properties
On the group level, we evaluated too many properties, because we would
pull in dependencies of dependencies, e.g. not only qbs.architecture,
but also cpp.qbs.architecture etc.
On the other hand, we also did not find all dependencies between
properties, because Export items were not properly handled.

The benchmarker tool run on qbs itself says:
========== Performance data for Resolving ==========
    Old instruction count: 2802829713
    New instruction count: 2431763918
    Relative change: -14 %
    Old peak memory usage: 22623344 Bytes
    New peak memory usage: 21095328 Bytes
    Relative change: -7 %

The real-world numbers for larger projects seem to be significantly
higher; with the Qt Creator super project, I am observing a 30% speed-up
on my machine.

Change-Id: Ifbff4a01ef6b5cf5801961a168b79a31d8c6b220
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>