Fix lookup of shadowed parent properties
Consider the following example

Project {
    name: "project"
    property string projectName: name
    Product {
        name: "product"
        property string foo: projectName
    }
}

The foo property is supposed to have the value "project".
Instead it had the value "product".

EvaluatorScriptClass::property() and SVConverter assume that data->item
is the item that's attached to the QScriptValue object. That was not the
case when a property was automatically looked up in the parent item.

Task-number: QBS-1117
Change-Id: I21d4cc4106ae631730cd34fe352459d311b09501
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>