Add a script to create the release packages for Windows
This is a simple script which invokes a build of Qbs using the
specific configuration we intend to use for creating official
Windows release packages. It only requires an existing qbs.exe in
the PATH as well as the QTDIR and QTDIR64 environment variables set
to 32-bit and 64-bit installations of Qt, respectively. It does not
use or modify the current environment outside of temporary
directories.

The intent is to use this script in conjunction with a Docker
container although that is not required; instructions for using
which will follow in a subsequent patch.

Change-Id: I3f150d53c5b4afffdb4d5f5f7d64bf322b263a78
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>