Fix handling of same subproject names in generated VS solutions
While Qbs product names must be unambiguous, using the same project
name several times is perfectly valid, as long as all projects with
the same name belong to a different root project.

Utilize the project's source file location to distinguish projects
that have the same name. Use this in conjunction with the name to
create an ID that is used as the key for the solutionFolders map.

Change-Id: I47079cd4207d505892e79755a3986614025d2749
Task-number: QBS-1082
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Thomas Epting <thomas.epting@stryker.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>