Provide more user-friendly errors in module validation
Instead of a cryptic message simply stating that various properties
are not set, try to provide the user with real-world, actionable
suggestions to aid in project setup.

This new style in the QNX module vs usage of the old
PropertyValidator is kind of a "trial" run and could be expanded to
the rest of the modules later.

Change-Id: I2763540351fa609a14470110a3971624a7f7cf13
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>