setup-qt: compress MSVC profiles even when Qt's MSVC version is unknown
"Compressing" MSVC profiles means that if MSVC2017-x64 and
MSVC2017-x86_x64 fully match, then we drop MSVC2017-x86_x64 and connect
the Qt profile to the former.

This compression took place only if there was a valid Qt version. Do the
compression always now, and take care that we throw out the right
cross-compiler toolchains.

While this patch does not fully fix QBS-1141, it alleviates the symptoms
if only one MSVC is installed, which is usually the case.

Task-number: QBS-1141
Change-Id: Ied766b6051074fbb92037185f2a357bcdbc55b68
Reviewed-by: Jake Petroules <jake.petroules@qt.io>