Remove the profile name from the unique product name
If the product is multiplexed by profile, its name is already uniquified
by the multiplex id. If the product is not multiplexed, we don't need to
make the name unique.
This removes references to the top-level profile name from the build
directory structure, which can be helpful when rebuilding with a
different profile name, but the same effective properties. Also, it
shortens the paths of generated artifacts in the common case.

Change-Id: I1586049b9650542732a64be4b7d5f2db4428d87e
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>