Refactor asset catalogs test for greater performance and coverage
This splits out assetCatalogsEmpty and assetCatalogsMultiple from the
main assetCatalog test. They are not affected by the data axis of the
main test, so this improves performance by not running each of them
twice for no reason. It also increases coverage by running them on
OSes < macOS 10.9 (which we still test with) as long as Xcode 5 is
present.

Further, the main assetCatalog test no longer has a hard requirement
on macOS 10.9 either, and instead conditionally enables testing of
asset catalogs based on the Xcode version (>= 5). This increases test
coverage of XIB compilation with older Xcodes and OSes. The MainMenu.xib
has been recreated in an old version of Xcode (4.6.3) so it is more
broadly usable as a test artifact with older configurations.

Tested on macOS 10.12.5 with Xcode 8.3.2, macOS 10.7.5 with Xcode 4.6.3,
and macOS 10.8.5 with Xcode 5.1.1.

Change-Id: Ifc5d1b0462fd94aa5f1d55727bc74af73e8b9272
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>