GCC: Try harder to find binutils binaries
Our basic assumption is that tools like nm, strip and so on are located
alongside the compiler. However, this is not always the case, e.g. when
using the icecream tool. Let's not annoy users of such tools by forcing
them to manually set properties, but look up the tools in the PATH if
they are not present alongside the compiler.

Change-Id: I4533dae8fe472eb82e6a4507e5c976e547ce80e7
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>