Do not create copies of precompiled header source files on MSVC
We used to create copies of cpp_pch_src and c_pch_src files in the MSVC
compiler rule to create an unambiguous file name that can be used in
later calls to the compiler rule.

This hack is not necessary anymore. We can inspect the artifacts of the
product in the compiler rule.

Change-Id: I40ff119e2507cae3210c52c82fd38cfd83cf261c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>