Remove Qt dependency from some tests in tst_api where it's easy to do so
These tests aren't testing actual "Qt" functionality, so don't let the
presence/absence of Qt influence the results.

The rest of the tests that depend on Qt actually *are* Qt-specific and
should go into their own test suite in a separate patch, or need to be
refactored/split in order to properly separate out the dependency.

Change-Id: I733ec8e9b0cc261f7976f97cab09fb21fe36cd34
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>