Move the compiler version detection into a separate Probe
The compiler version is needed in the GccProbe itself (to determine
which compiler flags are available to use) so it cannot be done in a
single probe. This change will be taken advantage of in a subsequent
patch.

This patch also fixes a silent failure that caused the wrong compiler
flags to be used in the GccProbe due to the compiler version being
undefined at the point where feature checks based on the compiler
version were being made... specifically, that the -target flag was
assumed to be unavailable due to the missing compiler version.

Change-Id: I13626c37f3fd21b5b8b6632714ba9dd9373b7838
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>