Give rules access to artifacts matching "explicitlyDependsOn"
If outputs "explicitly depend on" other artifacts, then it is very
likely that these are needed in the prepare script and/or the command.
In fact, such artifacts are almost like inputs, except that they are
always passed the same way, regardless of whether we have a multiplex
rule or not.

[ChangeLog] Artifacts corresponding to the "explicitlyDependsOn" property
are now available under this name in rules and commands.

Change-Id: I28e41c29e250cccb1a64e83152528c28f48d5703
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>