Cpp module: Default value for use*PrecompiledHeader is now true
This is of course the right default: If a precompiled header is present,
then you will tpyically want to use it. We just couldn't do that so far,
because we did not have access to the "explicitlyDependsOn" artifacts in
the compiler rule. Now we do.

[ChangeLog] The property cpp.useCxxPrecompiledHeader, as well as the
variants for the other languages, now default to true.

Change-Id: Ie04172790ad21b0ebdfbf3366b6f549f705e774c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>