Fix behavior with corrupt or outdated build graphs
Commit b6bf17cfd1 introduced a glitch in that a rebuild using a build
graph whose magic string does not match would throw an error at the
first attempt and silently succeed at the second one. The new behavior
is the same as for attempts to rebuild with property changes: A plain
"build" fails, a "resolve" does the trick. Some more code had to move
from the command line parser to the loader to achieve this.

Change-Id: I906748cc1686c0c5d413d9451ec9e2aa26ea2035
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>