Add the creation of a Chocolatey package to our release process
This is maintained by upstream developers and so should just be a part
of the official release process to begin with. This has been tested with
the Docker-based release instructions and provides the additional
security property to Chocolatey users of having an authoritative source
from which to obtain to the .nupkg which can then be installed without
network access to packages.chocolatey.org.

Change-Id: I6d20201451cd3a7f080a81c1dee6b515671e2105
Reviewed-by: Kai Koehne <kai.koehne@qt.io>