Don't force -undefined into linker flags
This is important if the user wants to use a different value than the
ones supported here. Also, change the default of
cpp.allowUnresolvedSymbols to undefined in accordance with the design
philosophy that 'undefined' typically results in no flags passed. This
will make no difference for user projects since the only platform
affected is Darwin and on that platform '-undefined error' is default.

Change-Id: I5557da37f4ed8c609e98a6f74ad4b553cae0b817
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>