Add reference pages to index.html
The reference is arguably the most important part, so it makes sense
to link also to sub-sections from the main page.

One of the issues is that the sub-pages are not properly 'linked'
with nextpage and previouspage. Anyhow, this should not block us
from making the index page more useful.

Change-Id: I1c80345795c1df294d07f7a31b613a24df341cb3
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>