Fix edge case in change tracking
If we have a moc'able header file whose moc output is included by the
cpp file, then it is an error not to list that header in the project
file, because moc will not run on it and the included moc_cpp file won't
be present, causing a compiler error. However, if the header was
originally listed in the project file and removed from it after
building, then the next attempt to build would still succeed, because
the generated artifact would only be removed from disk at the end of the
build process.
Fix this by amending the code that checks for removed child artifacts to
remove them from disk immediately, making sure a rebuild will not
accidentally succeed.

Change-Id: I4f69560cc899dbc61938acc1c697aa93407dc03c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>