Don't set the *_DEPLOYMENT_TARGET environment variables to empty string
This is a partial revert of a7247ff1cde96864243eba3bf5d36d6b5bded2d6,
which caused Qbs to set the *_DEPLOYMENT_TARGET environment variables to
the empty string if no value was provided for the deployment target.

The empty string causes problems for some tools like `strip`, which
passes the empty string verbatim to the -*_version_min flag of an
internal invocation of `ld`, which is not allowed.

Presumably the original intent was to influence the target platform that
the compiler would use, but this does not work (at least with current
toolchains) if the value is the empty string. It also shouldn't matter
since we use -target now, which explicitly states the platform. In that
case, the compiler will choose a minimum version internally if one is
not explicitly given.

Change-Id: I7d9f12bb52bf3289c0c5e3353a928c8cb0be9531
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>