Simplify qbs.hostOS
We remove several values that are obsolete, as well as values for
operating systems that have been discontinued.

bsd4 and qnx6 are removed because there is a separate qbs.hostOSVersion
property for the version of the host OS. These were removed from Qt
Creator in 2014.

osx is removed as it is a long-deprecated synonym for macos. 5 minor
releases since the introduction of 'macos' in Qbs 1.5.2 is long enough.

android, ios, lynx, nacl, qnx, integrity, and vxworks are removed
because it is either not possible or not practical to run Qbs on these
platforms. Command line tools can't be run on iOS or Android without
jailbreaking or rooting the device, QNX has long since removed the
QNX-native QCC compilers, and it seems unlikely that there have ever
been or would be compilers available for LynxOS, INTEGRITY or VxWorks
so those are easily out. NaCl is a deprecated technology that is being
replaced by WebAssembly (nor would you run a build process on either of
them anyways).

ultrix and reliant have not seen a release since 1995. dgux not since
2001. bsdi not since 2003. irix not since 2006. osf not since 2010.
dynix... unclear, but no later than 2002. symbian needs no mention.
All considered obsolete, all gone.

sco and unixware are effectively obsolete. Remove them until someone
expresses a real need.

haiku was added because Qt has it.

Change-Id: I5cc23ae165a3069c63ee13388655d0093e5ccabc
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>