Remove the need for calling TextFile.close()
It's very easy to forget to close an opened TextFile, and the widely
used boiler-plate finally blocks are not the most beautiful thing to
behold.

Remove the need for calling TextFile.close() by making sure we
automatically close files when we finish evaluating a chunk of
JavaScript.

[ChangeLog] TextFile objects are automatically closed after
the current script execution is finished.

Change-Id: I2bd8639951f4fa5ef797785952911accf33afded
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>